Add Bugsnag logging to API client
I've added some basic log calls on the API client, just to see how it works and what functionality we get from it. The plan in subsequent releases would be to go through the docs more thoroughly and see how we can fully implement logging in all the important areas.

I've also included a new settings module that changes the `BASE_URL` depending on the environment you are running in (dev, staging, prod). This comes with a new config file `eas.json` (if this doesn't work, I think it should go in `app.json`) somewhere. When publishing the app, we can then set the release channel to target different environments:

> `expo publish --release-channel prod`

for instance.